requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Login==0.6.3

# Database
SQLAlchemy==2.0.23

# Security
Werkzeug==3.0.1
PyJWT==2.8.0
bcrypt==4.1.2

# Validation and serialization
marshmallow==3.20.2
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Environment variables
python-dotenv==1.0.0
requests==2.32.3

# Type checking and development
mypy==1.7.1
types-Flask==1.1.6
types-Werkzeug==1.0.9

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Production server
gunicorn==21.2.0

# Development tools
black==23.11.0
flake8==6.1.0
isort==5.12.0